Added broken Redis proxy
Alas the Envoy redis proxy will not just pass through an AUTH request.
It wants to have the password defined in its configuration. I don't want
to back that out so maybe we can use a different strategy.

{"level":50,"time":1593364282143,"pid":17,"hostname":"shipit-v8-shipitclicker-6bd5cc7c7-c6zzl","name":"Shipit-Clicker-shipit-v8",  "msg":"Redis error","err":{"type":"ReplyError","message":"Ready check failed: invalid request","stack":"ReplyError: Ready check failed: invalid request\n    at parseError (/app/node_modules/redis-parser/lib/parser.js:193:12)\n    at parseType (/app/node_modules/redis-parser/lib/parser.js:303:14)","command":"INFO"},"v":1}

node_redis: Warning: Redis server does not require a password, but a password was supplied.